tests/wpt/mozilla/tests/mozilla/iframe_hierarchy.html

<!doctype html>
<meta charset="utf-8">
<title></title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe src="iframe_parent1.html"></iframe>
<script>
var t = async_test("iframe content hierarchy is correct");
function do_test() {
  t.step(function() {
    var doc = document.getElementsByTagName("iframe")[0].contentDocument;
    doc = doc.getElementsByTagName("iframe")[0].contentDocument;
    assert_equals(doc.body.firstChild.textContent, "foo\n");
    t.done();
  });
}
</script>
